refactor: move voxel stuff to engine
