+++ engine/src/cubos/engine/voxels/palette.cpp
@@ -1,9 +1,10 @@
 #include <cstring>
 
-#include <cubos/core/gl/palette.hpp>
 #include <cubos/core/log.hpp>
 
-using namespace cubos::core::gl;
+#include <cubos/engine/voxels/palette.hpp>
+
+using namespace cubos::engine;
 
 Palette::Palette(std::vector<Material>&& materials)
     : mMaterials(std::move(materials))
@@ -97,7 +98,7 @@ void Palette::merge(const Palette& palette, float similarity)
     }
 }
 
-void cubos::core::data::serialize(Serializer& serializer, const gl::Palette& palette, const char* name)
+void cubos::core::data::serialize(Serializer& serializer, const Palette& palette, const char* name)
 {
     // Count non-empty materials.
     std::size_t count = 0;
@@ -121,7 +122,7 @@ void cubos::core::data::serialize(Serializer& serializer, const gl::Palette& pal
     serializer.endDictionary();
 }
 
-void cubos::core::data::deserialize(Deserializer& deserializer, gl::Palette& palette)
+void cubos::core::data::deserialize(Deserializer& deserializer, Palette& palette)
 {
     palette.mMaterials.clear();
