+++ engine/src/cubos/engine/voxels/plugin.cpp
@@ -1,13 +1,10 @@
-#include <cubos/core/gl/grid.hpp>
-#include <cubos/core/gl/palette.hpp>
-
 #include <cubos/engine/assets/bridges/binary.hpp>
 #include <cubos/engine/assets/plugin.hpp>
+#include <cubos/engine/voxels/grid.hpp>
+#include <cubos/engine/voxels/palette.hpp>
 #include <cubos/engine/voxels/plugin.hpp>
 
 using cubos::core::ecs::Write;
-using cubos::core::gl::Grid;
-using cubos::core::gl::Palette;
 using namespace cubos::engine;
 
 static void bridges(Write<Assets> assets)
