+++ tools/quadrados/src/convert.cpp
@@ -5,13 +5,18 @@
 
 #include <cubos/core/data/binary_deserializer.hpp>
 #include <cubos/core/data/binary_serializer.hpp>
-#include <cubos/core/data/qb_parser.hpp>
-#include <cubos/core/gl/palette.hpp>
+#include <cubos/core/log.hpp>
+#include <cubos/core/memory/endianness.hpp>
 #include <cubos/core/memory/standard_stream.hpp>
+
+#include <cubos/engine/voxels/grid.hpp>
+#include <cubos/engine/voxels/palette.hpp>
 
 #include "tools.hpp"
 
-using namespace cubos::core;
+namespace data = cubos::core::data;
+namespace memory = cubos::core::memory;
+using namespace cubos::engine;
 
 namespace fs = std::filesystem;
 
@@ -26,6 +31,14 @@ struct ConvertOptions
     bool force = false;                              ///< Enables force mode.
     bool help = false;                               ///< Prints the help message.
     float similarity = 1.0F;                         ///< The similarity threshold.
+};
+
+/// @brief Represents the data read from a matrix in a QB file.
+struct QBMatrix
+{
+    Grid grid;           ///< Grid of the matrix.
+    Palette palette;     ///< Palette of the matrix.
+    glm::ivec3 position; ///< Position of the matrix.
 };
 
 /// Prints the help message of the program.
@@ -154,10 +167,160 @@ static bool parseArguments(int argc, char** argv, ConvertOptions& options)
     return true;
 }
 
+/// @brief Parses a Qubicle file (.qb), pushing every matrix found in the file to the passed vector.
+/// @param[out] matrices Parsed matrices.
+/// @param stream Stream to read from.
+/// @return Whether the file was parsed successfully.
+bool parseQB(std::vector<QBMatrix>& matrices, memory::Stream& stream)
+{
+    uint8_t version[4] = {0, 0, 0, 0};
+    uint32_t colorFormat;
+    uint32_t zAxisOrientation;
+    uint32_t compressed;
+    uint32_t visibilityMaskEncoded;
+    uint32_t numMatrices;
+
+    // Parse the version of the file.
+    stream.read(version, 4);
+    if (stream.eof())
+    {
+        CUBOS_ERROR("Unexpected end of file while reading file version");
+        return false;
+    }
+    if (version[0] != 1 || version[1] != 1 || version[2] != 0 || version[3] != 0)
+    {
+        CUBOS_ERROR("Unsupported version {}.{}.{}.{}, only 1.1.0.0 is supported", version[0], version[1], version[2],
+                    version[3]);
+        return false;
+    }
+
+    // Parse the rest of the header.
+    stream.read(&colorFormat, 4);
+    stream.read(&zAxisOrientation, 4);
+    stream.read(&compressed, 4);
+    stream.read(&visibilityMaskEncoded, 4);
+    stream.read(&numMatrices, 4);
+    colorFormat = memory::fromLittleEndian(colorFormat);
+    zAxisOrientation = memory::fromLittleEndian(zAxisOrientation);
+    compressed = memory::fromLittleEndian(compressed);
+    visibilityMaskEncoded = memory::fromLittleEndian(visibilityMaskEncoded);
+    numMatrices = memory::fromLittleEndian(numMatrices);
+
+    if (stream.eof())
+    {
+        CUBOS_ERROR("Unexpected end of file while reading file header");
+        return false;
+    }
+    if (visibilityMaskEncoded != 0U)
+    {
+        CUBOS_ERROR("Visibility mask encoding is not supported");
+        return false;
+    }
+
+    // Parse the matrices.
+    matrices.resize(numMatrices);
+    for (uint32_t i = 0; i < numMatrices; ++i)
+    {
+        uint8_t nameLen;
+        uint32_t sizeX;
+        uint32_t sizeY;
+        uint32_t sizeZ;
+        int32_t posX;
+        int32_t posY;
+        int32_t posZ;
+
+        // Read the matrix name.
+        stream.read(&nameLen, 1);
+        std::string name(nameLen, ' ');
+        stream.read(name.data(), nameLen);
+
+        // Read the matrix size.
+        stream.read(&sizeX, 4);
+        stream.read(&sizeY, 4);
+        stream.read(&sizeZ, 4);
+        sizeX = memory::fromLittleEndian(sizeX);
+        sizeY = memory::fromLittleEndian(sizeY);
+        sizeZ = memory::fromLittleEndian(sizeZ);
+        matrices[i].grid.setSize({sizeX, sizeY, sizeZ});
+
+        // Read the matrix position.
+        stream.read(&posX, 4);
+        stream.read(&posY, 4);
+        stream.read(&posZ, 4);
+        posX = memory::fromLittleEndian(posX);
+        posY = memory::fromLittleEndian(posY);
+        posZ = memory::fromLittleEndian(posZ);
+        matrices[i].position = glm::ivec3(posX, posY, posZ);
+
+        // Read the matrix voxels.
+        if (compressed == 0)
+        {
+            uint8_t color[4];
+            std::size_t nextMat = 1;
+
+            for (uint32_t z = 0; z < sizeZ; ++z)
+            {
+                for (uint32_t y = 0; y < sizeY; ++y)
+                {
+                    for (uint32_t x = 0; x < sizeX; ++x)
+                    {
+                        stream.read(color, 4);
+                        if (color[3] == 0)
+                        {
+                            continue;
+                        }
+                        if (colorFormat != 0U)
+                        { // BGRA -> RGBA
+                            std::swap(color[0], color[2]);
+                        }
+                        glm::vec4 colorVec(static_cast<float>(color[0]) / 255.0F, static_cast<float>(color[1]) / 255.0F,
+                                           static_cast<float>(color[2]) / 255.0F,
+                                           static_cast<float>(color[3]) / 255.0F);
+
+                        // Check if the material is already in the palette.
+                        std::size_t mat;
+                        for (mat = 1; mat < nextMat; ++mat)
+                        {
+                            if (matrices[i].palette.get(static_cast<uint16_t>(mat)).color == colorVec)
+                            {
+                                break;
+                            }
+                        }
+                        if (mat == nextMat)
+                        {
+                            if (mat >= 65536)
+                            {
+                                CUBOS_ERROR("Too many materials, max is 65536");
+                                return false;
+                            }
+
+                            // Add the material to the palette.
+                            Material desc;
+                            desc.color = colorVec;
+                            matrices[i].palette.set(static_cast<uint16_t>(mat), desc);
+                            nextMat += 1;
+                        }
+
+                        // Set the voxel.
+                        matrices[i].grid.set(glm::ivec3(x, y, z), static_cast<uint16_t>(mat));
+                    }
+                }
+            }
+        }
+        else
+        {
+            CUBOS_ERROR("Compressed QB files are not supported");
+            return false;
+        }
+    }
+
+    return true;
+}
+
 /// Tries to load the palette from the given path.
 /// @param path The path of the palette.
 /// @param palette The palette to fill.
-static bool loadPalette(const fs::path& path, gl::Palette& palette)
+static bool loadPalette(const fs::path& path, Palette& palette)
 {
     auto* file = fopen(path.string().c_str(), "rb");
     if (file == nullptr)
@@ -180,7 +343,7 @@ static bool loadPalette(const fs::path& path, gl::Palette& palette)
 /// Tries to load a QBModel from the given path.
 /// @param path The path of the QBModel.
 /// @param model The model to fill.
-static bool loadQB(const fs::path& path, std::vector<data::QBMatrix>& model)
+static bool loadQB(const fs::path& path, std::vector<QBMatrix>& model)
 {
     auto* file = fopen(path.string().c_str(), "rb");
     if (file == nullptr)
@@ -189,13 +352,13 @@ static bool loadQB(const fs::path& path, std::vector<data::QBMatrix>& model)
     }
 
     auto stream = memory::StandardStream(file, true);
-    return data::parseQB(model, stream);
+    return parseQB(model, stream);
 }
 
 /// Saves the given palette to the given path.
 /// @param path The path of the palette.
 /// @param palette The palette to save.
-static bool savePalette(const fs::path& path, const gl::Palette& palette)
+static bool savePalette(const fs::path& path, const Palette& palette)
 {
     auto* file = fopen(path.string().c_str(), "wb");
     if (file == nullptr)
@@ -218,7 +381,7 @@ static bool savePalette(const fs::path& path, const gl::Palette& palette)
 /// Saves the given grid to the given path.
 /// @param path The path of the grid.
 /// @param grid The grid to export.
-static bool saveGrid(const fs::path& path, const gl::Grid& grid)
+static bool saveGrid(const fs::path& path, const Grid& grid)
 {
     auto* file = fopen(path.string().c_str(), "wb");
     if (file == nullptr)
@@ -244,7 +407,7 @@ static bool saveGrid(const fs::path& path, const gl::Grid& grid)
 static bool convert(const ConvertOptions& options)
 {
     // First, load the palette.
-    gl::Palette palette;
+    Palette palette;
     if (!options.palette.empty())
     {
         if (!loadPalette(options.palette, palette))
@@ -268,7 +431,7 @@ static bool convert(const ConvertOptions& options)
         std::cerr << "Failed to load model: unsupported format " << options.input.extension() << "." << std::endl;
     }
 
-    std::vector<data::QBMatrix> model;
+    std::vector<QBMatrix> model;
     if (!loadQB(options.input, model))
     {
         std::cerr << "Failed to load model: " << options.input << " not found." << std::endl;
